+++ src/views/toDoModal/repeatingEvent.vue
@@ -1,5 +1,6 @@
 <template>
-  <div id="reDropDown" class="header-menu-icons" type="button" data-bs-toggle="dropdown" data-bs-auto-close="outside" :title="$t('ui.recurringTasks')">
+  <div id="reDropDown" class="header-menu-icons" type="button" data-bs-toggle="dropdown" data-bs-auto-close="outside"
+    :title="$t('ui.recurringTasks')">
     <i id="btnRepeatingEvent" :class="{ 'bi-arrow-clockwise': !repeatingEvent, 'bi-arrow-repeat': repeatingEvent }"></i>
   </div>
 
@@ -9,17 +10,37 @@
         :disabled="repeatingEvent">
         <option value="">{{ $t("todoDetails.noRepeat") }}</option>
         <option value="3">{{ $t("todoDetails.daily") }}</option>
-        <option value="4">{{ $t("todoDetails.weekdays") }} </option>
         <option value="2">{{ $t("todoDetails.weekly") }} </option>
+        <option value="4">{{ $t("todoDetails.weekdays") }} </option>
+        <option value="5"> {{ $t("todoDetails.customWeekdays") }} </option>
         <option value="1">{{ $t("todoDetails.monthly") }}</option>
         <option value="0">{{ $t("todoDetails.yearly") }}</option>
       </select>
 
       <div v-if="repeatingType" class="d-flex flex-column">
+
+        <div v-show="repeatingType == '5' || (repeatingEvent && repeatingType == '2')"
+          class="weekDays-selector mt-3 mb-1">
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-mon" class="weekday" v-model="weekdays.mon" />
+          <label for="weekday-mon">{{ moments().isoWeekday(1).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-tue" class="weekday" v-model="weekdays.tue" />
+          <label for="weekday-tue">{{ moments().isoWeekday(2).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-wed" class="weekday" v-model="weekdays.wed" />
+          <label for="weekday-wed">{{ moments().isoWeekday(3).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-thu" class="weekday" v-model="weekdays.thu" />
+          <label for="weekday-thu">{{ moments().isoWeekday(4).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-fri" class="weekday" v-model="weekdays.fri" />
+          <label for="weekday-fri">{{ moments().isoWeekday(5).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-sat" class="weekday" v-model="weekdays.sat" />
+          <label for="weekday-sat">{{ moments().isoWeekday(6).locale(language).format("dd")[0] }}</label>
+          <input type="checkbox" :disabled="repeatingEvent" id="weekday-sun" class="weekday" v-model="weekdays.sun" />
+          <label for="weekday-sun">{{ moments().isoWeekday(7).locale(language).format("dd")[0] }}</label>
+        </div>
+
+
         <div class="d-flex align-items-center justify-content-between re-input">
           <label class="opacity-50">{{ $t("todoDetails.interval") }}</label>
-          <input type="number" class="form-control lex-shrink-1 counter" v-model="interval"
-            :disabled="repeatingEvent" />
+          <input type="number" class="form-control lex-shrink-1 counter" v-model="interval" :disabled="repeatingEvent" />
         </div>
 
         <select class="form-select re-input" aria-label="Default select example" v-model="ocurrencesType"
@@ -67,6 +88,7 @@ export default {
       ocurrencesType: "",
       ocurrences: 1,
       untilDate: null,
+      weekdays: { mon: false, tue: false, wed: false, thu: false, fri: false, sat: false, sun: false },
     };
   },
   props: {
@@ -76,7 +98,6 @@ export default {
   methods: {
     done() {
       const rule = this.repeatingEventRule();
-      console.log(rule);
       var repeatingEventId = this.repeatingEvent ? this.repeatingEvent : moment().format("x");
       if (rule) {
         let date = this.todo.listId;
@@ -112,20 +133,28 @@ export default {
     repeatingEventRule() {
       if (!this.repeatingType) return null;
 
-       var ruleOptions = {
+      var ruleOptions = {
         freq: this.repeatingType,
         interval: this.interval,
         dtstart: moment.utc(this.todo.listId, "YYYYMMDD").toDate(),
       };
 
-      console.log(moment(this.todo.listId, "YYYYMMDD").toDate())
-      console.log(moment.utc(this.todo.listId, "YYYYMMDD").toDate())
-
-      if (this.repeatingType == 4){
+      if (this.repeatingType == 4) {
         ruleOptions.byweekday = [RRule.MO, RRule.TU, RRule.WE, RRule.TH, RRule.FR];
         ruleOptions.freq = 2;
       }
-      
+
+      if (this.repeatingType == 5) {
+        ruleOptions.byweekday = [];
+        this.weekdays.mon && ruleOptions.byweekday.push(RRule.MO);
+        this.weekdays.tue && ruleOptions.byweekday.push(RRule.TU);
+        this.weekdays.wed && ruleOptions.byweekday.push(RRule.WE);
+        this.weekdays.thu && ruleOptions.byweekday.push(RRule.TH);
+        this.weekdays.fri && ruleOptions.byweekday.push(RRule.FR);
+        this.weekdays.sat && ruleOptions.byweekday.push(RRule.SA);
+        this.weekdays.sun && ruleOptions.byweekday.push(RRule.SU);
+        ruleOptions.freq = 2;
+      }
 
       if (this.ocurrencesType == "ocurrences") {
         ruleOptions.count = this.ocurrences;
@@ -159,10 +188,14 @@ export default {
 
       return re_event;
     },
+    moments: function (date) {
+      return moment(date);
+    }
   },
   watch: {
     repeatingEvent: function (newVal) {
       let re = this.$store.getters.repeatingEventList[newVal];
+      this.weekdays = { mon: false, tue: false, wed: false, thu: false, fri: false, sat: false, sun: false };
       if (re) {
         const rule = rrulestr(re.repeating_rule);
         this.repeatingType = rule.options.freq;
@@ -170,6 +203,16 @@ export default {
         this.ocurrences = rule.options.count;
         this.ocurrencesType = re.ocurrencesType;
         this.untilDate = rule.options.until ? rule.options.until.toLocaleDateString("en-GB").split("/").reverse().join("-") : null;
+        if (rule.options.byweekday) {
+          rule.options.byweekday.includes(0) && (this.weekdays.mon = true);
+          rule.options.byweekday.includes(1) && (this.weekdays.tue = true);
+          rule.options.byweekday.includes(2) && (this.weekdays.wed = true);
+          rule.options.byweekday.includes(3) && (this.weekdays.thu = true);
+          rule.options.byweekday.includes(4) && (this.weekdays.fri = true);
+          rule.options.byweekday.includes(5) && (this.weekdays.sat = true);
+          rule.options.byweekday.includes(6) && (this.weekdays.sun = true);
+        }
+
       } else {
         this.repeatingType = "";
         this.ocurrencesType = "";
@@ -180,6 +223,11 @@ export default {
       }
     },
   },
+  computed: {
+    language: function () {
+      return this.$store.getters.config.language;
+    },
+  }
 };
 </script>
 
@@ -233,5 +281,42 @@ export default {
   background-color: #21262d;
   color: #c9d1d9;
   border-bottom: 1px solid #464647;
+}
+
+.weekDays-selector {
+  display: flex;
+  justify-content: center;
+}
+
+.weekDays-selector input {
+  display: none !important;
+}
+
+.weekDays-selector input[type=checkbox]+label {
+  display: inline-block;
+  border-radius: 3px;
+  background: #eaecef;
+  height: 20px;
+  width: 20px;
+  margin-right: 3px;
+  line-height: 20px;
+  text-align: center;
+  cursor: pointer;
+  font-size: 12px;
+}
+
+.dark-theme .weekDays-selector input[type=checkbox]+label {
+  background: #15161e;
+  color: #ffffff;
+}
+
+.weekDays-selector input[type=checkbox]:disabled+label {
+  cursor: unset;
+  opacity: 0.6;
+}
+
+.weekDays-selector input[type=checkbox]:checked+label {
+  background: #5c5c5c;
+  color: #ffffff;
 }
 </style>